+++ CHANGELOG.md
@@ -10,7 +10,8 @@
 * Allow caching costlocker responses with `--cache` option
     * `--cache` - in verbose mode (`-vvv`) it also prints if costlocker is called or cached response is used
     * `bin/clear-cache` - clear responses cache
-
+* Export
+    * Mailer refactoring - move to `Costlocker\Reports\Export`, catch `Swift_Mailer` exceptions, don't unlink sent file
 
 ## v1.0.1 (_2017-04-10_)